% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spidR.R
\name{records}
\alias{records}
\title{Downloads coordinate data.}
\usage{
records(sp)
}
\arguments{
\item{sp}{A single species name.}
}
\value{
A data.frame with longitude and latitude.
}
\description{
Downloads coordinate data from GBIF and the World Spider Trait Database.
}
\details{
Outputs non-duplicate records with geographical (long, lat) coordinates. As always when using data from multiple sources the user should be careful and check if records "make sense" before using them.
}
\examples{
\dontrun{
records("Pardosa hyperborea")
}
}
\references{
Pekar, S., Cernecka, L., Wolff, J., Mammola, S., Cardoso, P., Lowe, E., Fukushima, C.S., Birkhofer, K. & Herberstein, M.E. (2021). The spider trait database. Masaryk University, Brno, URL: https://spidertraits.sci.muni.cz
}
