% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spidR.R
\name{species}
\alias{species}
\title{Get species from higher taxa.}
\usage{
species(tax, order = FALSE)
}
\arguments{
\item{tax}{A taxon name or vector with taxa names.}

\item{order}{Order species names alphabetically.}
}
\value{
A vector with species names.
}
\description{
Get species within given families or genera from the World Spider Catalogue.
}
\details{
This function will get all species currently listed for given families or genera from the World Spider Catalogue (2021).
}
\examples{
\dontrun{
species("Amphiledorus")
species(tax = c("Amphiledorus", "Nemesiidae"), order = TRUE)
}
}
\references{
World Spider Catalog (2021). World Spider Catalog. Version 22.0. Natural History Museum Bern, online at http://wsc.nmbe.ch. doi: 10.24436/2.
}
