% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spidR.R
\name{wsc.data}
\alias{wsc.data}
\title{Downloads WSC data.}
\usage{
wsc.data()
}
\value{
A matrix with all current species names and distribution. This should be used for other functions using wsc data.
}
\description{
Downloads the most recent data from the World Spider Catalogue.
}
\details{
The World Spider Catalog (2021) lists all currently valid species of spiders, from Clerck to date. Updated daily.
}
\examples{
\dontrun{
wsc.data()
}
}
\references{
World Spider Catalog (2021). World Spider Catalog. Version 22.0. Natural History Museum Bern, online at http://wsc.nmbe.ch. doi: 10.24436/2.
}
