% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spidR.R
\docType{data}
\name{wscmap}
\alias{wscmap}
\title{Distribution of spiders at the World Spider Catalogue.}
\format{
A matrix with regions and corresponding ISO3 codes.
}
\usage{
data(wscmap)
}
\description{
A dataset that links species distribution descriptions with the map using the ISO3 code
}
\keyword{datasets}
